embedding/requirements.txt
fastapi
uvicorn
faiss-cpu
llama-index
openai
python-dotenv
llama-index-vector-stores-faiss
boto3
psycopg2